Move clients into own namespace, rename interfaces
